Add ResourceLoader#exists (Android/JVM)
