+++ util-test/src/commonMain/kotlin/care/data4life/sdk/util/test/Constants.kt
@@ -16,28 +16,6 @@
 
 package care.data4life.sdk.util.test
 
-import java.io.File
-import care.data4life.sdk.util.test.lang.FileNotFoundError
-
-actual class CommonResourceLoader actual constructor(projectDir: AbsolutePath) {
-    private val projectPath = projectDir
-
-    private fun resolveRoot(root: Path?): Path {
-        return if (root is Path) {
-            root.trimEnd('/')
-        } else {
-            Constants.commonRoot
-        }
-    }
-
-    actual fun exists(path: Path, root: Path?): Boolean {
-        val resource = File(
-            "${resolveRoot(root)}/${path.trimStart('/')}"
-        )
-
-        return resource.exists()
-    }
-
-    @Throws(FileNotFoundError::class)
-    actual fun load(path: Path, root: Path?): String = TODO()
+internal object Constants {
+    const val commonRoot: Path = "src/commonTest/resources"
 }
