+++ util-test/src/jvmMain/kotlin/care/data4life/sdk/util/test/CommonResourceLoader.kt
@@ -16,11 +16,27 @@
 
 package care.data4life.sdk.util.test
 
+import java.io.File
 import care.data4life.sdk.util.test.lang.FileNotFoundError
 
 actual class CommonResourceLoader actual constructor(projectDir: AbsolutePath) {
+    private val projectPath = projectDir
 
-    actual fun exists(path: Path, root: Path?): Boolean = TODO()
+    private fun resolveRoot(root: Path?): Path {
+        return if (root is Path) {
+            root.trimEnd('/')
+        } else {
+            Constants.commonRoot
+        }
+    }
+
+    actual fun exists(path: Path, root: Path?): Boolean {
+        val resource = File(
+            "${resolveRoot(root)}/${path.trimStart('/')}"
+        )
+
+        return resource.exists()
+    }
 
     @Throws(FileNotFoundError::class)
     actual fun load(path: Path, root: Path?): String = TODO()
